refactor: split suites page into components
