+++ lib/static/new-ui/app/report.tsx
@@ -1,4 +1,4 @@
-import React, {useEffect} from 'react';
+import React, {ReactNode, useEffect} from 'react';
 import {createRoot} from 'react-dom/client';
 import {App} from './App';
 import store from '../../modules/store';
@@ -7,7 +7,7 @@ import {initStaticReport, finStaticReport} from '../../modules/actions';
 const rootEl = document.getElementById('app') as HTMLDivElement;
 const root = createRoot(rootEl);
 
-function Gui(): React.JSX.Element {
+function Gui(): ReactNode {
     useEffect(() => {
         store.dispatch(initStaticReport());
