+++ lib/static/new-ui/features/suites/components/SuitesPage/index.tsx
@@ -1,184 +1,25 @@
-import React, {ChangeEvent, useCallback, useEffect, useState} from 'react';
-import {Box, Flex, TextInput} from '@gravity-ui/uikit';
-import {debounce} from 'lodash';
-import classNames from 'classnames';
+import {Flex} from '@gravity-ui/uikit';
+import React, {ReactNode} from 'react';
 import {connect} from 'react-redux';
-import {
-    unstable_useList as useList,
-    unstable_ListContainerView as ListContainerView,
-    unstable_ListItemView as ListItemView,
-    unstable_getItemRenderState as getItemRenderState
-} from '@gravity-ui/uikit/unstable';
 
-import styles from './index.module.css';
-import {useVirtualizer} from '@tanstack/react-virtual';
-import {bindActionCreators} from 'redux';
-import {State} from '@/static/new-ui/types/store';
-import {
-    getTreeViewExpandedById,
-    getTreeViewItems
-} from '@/static/new-ui/features/suites/components/SuitesPage/selectors';
-import {
-    TreeViewBrowserData,
-    TreeViewItem, TreeViewItemType,
-    TreeViewSuiteData
-} from '@/static/new-ui/features/suites/components/SuitesPage/types';
 import {SplitViewLayout} from '@/static/new-ui/components/SplitViewLayout';
-import {getIconByStatus} from '@/static/new-ui/utils';
-import {TreeViewItemTitle} from '@/static/new-ui/features/suites/components/TreeViewItemTitle';
-import {TreeViewItemSubtitle} from '@/static/new-ui/features/suites/components/TreeViewItemSubtitle';
-import * as actions from '@/static/modules/actions';
-import {useNavigate, useParams} from 'react-router-dom';
+import {TestNameFilter} from '@/static/new-ui/features/suites/components/TestNameFilter';
+import {SuitesTreeView} from '@/static/new-ui/features/suites/components/SuitesTreeView';
+import styles from './index.module.css';
 
-interface SuitesPageInternalProps {
-    treeViewItems: TreeViewItem<TreeViewSuiteData | TreeViewBrowserData>[];
-    treeViewExpandedById: Record<string, boolean>;
-    actions: typeof actions;
-    isInitialized: boolean;
-    currentSuiteId: string | null;
-    testNameFilter: string;
-}
-
-function SuitesPageInternal(props: SuitesPageInternalProps): JSX.Element {
-    const navigate = useNavigate();
-    const {suiteId} = useParams();
-    const itemsOriginal = props.treeViewItems;
-
-    const list = useList({
-        items: itemsOriginal,
-        withExpandedState: true,
-        getItemId: item => {
-            return item.fullTitle;
-        },
-        controlledState: {
-            expandedById: props.treeViewExpandedById
-        }
-    });
-
-    const onItemClick = useCallback(({id}: {id: string}): void => {
-        const item = list.structure.itemsById[id];
-
-        if (item.type === TreeViewItemType.Suite && list.state.expandedById && id in list.state.expandedById && list.state.setExpanded) {
-            props.actions.toggleSuiteSection({suiteId: item.fullTitle, shouldBeOpened: !props.treeViewExpandedById[item.fullTitle]});
-        } else if (item.type === TreeViewItemType.Browser) {
-            props.actions.suitesPageSetCurrentSuite(id);
-
-            navigate(encodeURIComponent(item.fullTitle));
-        }
-    }, [list, props.actions, props.treeViewExpandedById]);
-
-    const updateTestNameFilter = useCallback(debounce(
-        (testName) => props.actions.updateTestNameFilter(testName),
-        500,
-        {maxWait: 3000}
-    ), []);
-
-    const parentRef = React.useRef<HTMLDivElement>(null);
-
-    const virtualizer = useVirtualizer({
-        count: list.structure.visibleFlattenIds.length,
-        getScrollElement: () => parentRef.current,
-        estimateSize: () => 28,
-        getItemKey: useCallback((index: number) => list.structure.visibleFlattenIds[index], [list]),
-        overscan: 200
-    });
-
-    useEffect(() => {
-        if (!suiteId && props.currentSuiteId && suiteId !== props.currentSuiteId) {
-            navigate(encodeURIComponent(props.currentSuiteId));
-        }
-    }, []);
-
-    useEffect(() => {
-        if (!props.isInitialized) {
-            return;
-        }
-
-        props.actions.setStrictMatchFilter(false);
-
-        if (suiteId) {
-            props.actions.suitesPageSetCurrentSuite(suiteId);
-            virtualizer.scrollToIndex(list.structure.visibleFlattenIds.indexOf(suiteId), {align: 'start'});
-        }
-    }, [props.isInitialized]);
-
-    const [testNameFilter, setTestNameFilter] = useState(props.testNameFilter);
-    const onChange = useCallback((event: ChangeEvent<HTMLInputElement>): void => {
-        setTestNameFilter(event.target.value);
-        updateTestNameFilter(event.target.value);
-    }, [setTestNameFilter, updateTestNameFilter]);
-
-    const items = virtualizer.getVirtualItems();
-
+function SuitesPageInternal(): ReactNode {
     return <SplitViewLayout>
         <div>
             <Flex direction={'column'} spacing={{p: '2'}} style={{height: '100vh'}}>
                 <h2 className="text-display-1">Suites</h2>
                 <div className={styles.controlsRow}>
-                    <TextInput placeholder='Search or filter' value={testNameFilter} onChange={onChange}/>
+                    <TestNameFilter/>
                 </div>
-                <ListContainerView className={styles.treeView}>
-                    <div ref={parentRef} className={styles['tree-view__container']}>
-                        <div
-                            className={styles['tree-view__total-size-wrapper']}
-                            style={{height: virtualizer.getTotalSize()}}
-                        >
-                            <div
-                                className={styles['tree-view__visible-window']}
-                                style={{transform: `translateY(${items[0]?.start ?? 0}px)`}}
-                            >
-                                {items.map((virtualRow) => {
-                                    const item = list.structure.itemsById[virtualRow.key];
-                                    const classes = [styles['tree-view__item']];
-                                    if (item.fullTitle === props.currentSuiteId) {
-                                        classes.push(styles['tree-item--current']);
-                                    } else if ((item.status === 'fail' || item.status === 'error') && item.type === TreeViewItemType.Browser) {
-                                        classes.push(styles['tree-item--error']);
-                                    }
-                                    if (item.type === TreeViewItemType.Browser) {
-                                        classes.push(styles['tree-item--browser']);
-                                    }
-
-                                    return <Box
-                                        key={virtualRow.key}
-                                        data-index={virtualRow.index}
-                                        ref={virtualizer.measureElement}
-                                        spacing={{pt: 1}}
-                                    >
-                                        <ListItemView
-                                            height={0}
-                                            className={classNames(classes)}
-                                            {...getItemRenderState({
-                                                id: virtualRow.key.toString(),
-                                                list,
-                                                onItemClick,
-                                                mapItemDataToContentProps: (x) => {
-                                                    return {
-                                                        startSlot: getIconByStatus(x.status),
-                                                        title: <TreeViewItemTitle item={x}/>,
-                                                        subtitle: <TreeViewItemSubtitle item={x}/>
-                                                    };
-                                                }
-                                            }).props}/>
-                                    </Box>;
-                                })}
-                            </div>
-                        </div>
-                    </div>
-                </ListContainerView>
+                <SuitesTreeView/>
             </Flex>
         </div>
         <div></div>
     </SplitViewLayout>;
 }
 
-export const SuitesPage = connect(
-    (state: State) => ({
-        isInitialized: state.app.isInitialized,
-        currentSuiteId: state.app.currentSuiteId,
-        treeViewItems: getTreeViewItems(state),
-        treeViewExpandedById: getTreeViewExpandedById(state),
-        testNameFilter: state.view.testNameFilter
-    }),
-    (dispatch) => ({actions: bindActionCreators(actions, dispatch)})
-)(SuitesPageInternal);
+export const SuitesPage = connect()(SuitesPageInternal);
